Remove usage of unit type in Converter's NIF

+++ c_src/membrane_ffmpeg_swscale_plugin/pixel_format_converter/converter.spec.exs
@@ -1,7 +1,7 @@
 module Membrane.FFmpeg.SWScale.PixelFormatConverter.Native
 state_type "State"
 
-spec create( width :: uint, height :: uint, old_format :: atom, new_format :: atom) :: {:ok :: label, state} | {:error :: label, reason :: atom}
+spec create(width :: uint64, height :: uint64, old_format :: atom, new_format :: atom) :: {:ok :: label, state} | {:error :: label, reason :: atom}
 
 spec process(state, payload) :: {:ok :: label, payload} | {:error :: label, reason :: atom}
